+++ lib/bashcov/lexer.rb
@@ -1,11 +1,13 @@
 # frozen_string_literal: true
 
-require "bashcov/line"
+require "shellwords"
 
 module Bashcov
   # Simple lexer which analyzes Bash files in order to get information for
   # coverage
-  class Lexer
+  module Lexer
+    IGNORE_COMMENT = %r/\s*#/
+
     # Lines starting with one of these tokens are irrelevant for coverage
     IGNORE_START_WITH = %w(# function).freeze
 
@@ -15,35 +17,14 @@ module Bashcov
     # Lines containing only one of these keywords are irrelevant for coverage
     IGNORE_IS = %w(esac if then else elif fi while do done { } ;;).freeze
 
-    # @param [String] filename File to analyze
-    # @param [Hash] coverage Coverage with executed lines marked
-    # @raise [ArgumentError] if the given +filename+ is invalid.
-    def initialize(filename, coverage)
-      @filename = filename
-      @coverage = coverage
-
-      unless File.file?(@filename)
-        raise ArgumentError, "#{@filename} is not a file"
-      end
+  module_function
+    def comment?(l)
+      IGNORE_COMMENT =~ l.strip
     end
 
-    # Yields uncovered relevant lines.
-    # @note Uses +@coverage+ to avoid wasting time parsing executed lines.
-    # @return [void]
-    def uncovered_relevant_lines
-      lineno = 0
-      File.open(@filename, "rb").each_line do |line|
-        if @coverage[lineno] == Bashcov::Line::IGNORED && relevant?(line)
-          yield lineno
-        end
-        lineno += 1
-      end
-    end
-
-  private
-
-    def relevant?(line)
-      line.strip!
+    def relevant?(l)
+      # +l+ is can be frozen
+      line = l.strip
 
       !line.empty? and
         !IGNORE_IS.include? line and